changed array size to long, Closes #114

+++ src/symbol_table.c
@@ -88,7 +88,7 @@ mcc_symbol_table_new_row_function(char *name, enum mcc_symbol_table_row_type typ
 }
 
 struct mcc_symbol_table_row *mcc_symbol_table_new_row_array(char *name,
-                                                            int array_size,
+                                                            long array_size,
                                                             enum mcc_symbol_table_row_type type,
                                                             struct mcc_ast_node *node)
 {
